Refactor news management and update related components
- Updated GalleryController, GuruController, KurikulumController, and MuridController to handle Cloudinary image uploads more robustly.
- Introduced StoreRequest and UpdateRequest for news validation.
- Refactored NewsController to use new request classes and improved error handling.
- Created a new Berita model and migration for the berita table.
- Updated views for news and curriculum management to reflect new data structure and improve user experience.
- Added pagination and search functionality for news.
- Removed obsolete news migration file.

+++ sekolah/app/Http/Controllers/Admin/KurikulumController.php
@@ -96,9 +96,12 @@ class KurikulumController extends Controller
 
         $result = $kurikulum->update($data);
         if (!$result) {
-            cloudinary()->uploadApi()->destroy($file['public_id'], [
-                'resource_type' => 'raw'
-            ]);
+            if (isset($file) && $file['public_id']) {
+                // If the file upload failed, remove the uploaded file from Cloudinary
+                cloudinary()->uploadApi()->destroy($file['public_id'], [
+                    'resource_type' => 'raw'
+                ]);
+            }
             return redirect()->back()->with('error', 'Gagal memperbarui data Kalender Pendidikan');
         }
